JAGs, Light Ring, and Shooting
The JAGs work with the vacuum and the current can be read. Arm movement
revised. Light Ring revised but not finished. Still needs an upper
limit. Shoot timing can be changed using the axis-4 on the silver
joystick and the voltage for the vacuum rises to 1 when the current
drops.